+++ homework-5/src/main/java/by/andrei/firstproject/homework_5/data/CarDAO.kt
@@ -8,7 +8,7 @@ interface CarDAO {
     fun getCarsList(): MutableList<Car>
 
     @Query("SELECT * FROM car WHERE id = :carId")
-    fun getCar(carId: Long): Car
+    fun getCar(carId: Int): Car
 
     @Insert
     fun insertAll(car: Car)
